pyproject.toml
[tool.poetry]
name = "exams-api"
version = "0.1.0"
description = ""
authors = ["Your Name <you@example.com>"]
readme = "README.md"
packages = [{include = "app"}]

[tool.poetry.dependencies]
python = "^3.11"
fastapi = {extras = ["all"], version = "^0.100.0"}
psycopg = {extras = ["binary", "pool"], version = "^3.1.8"}
sqlalchemy = "^2.0.9"
alembic = "^1.10.3"
pydantic-settings = "^2.0.2"

[tool.poetry.group.dev.dependencies]
pytest = "^7.3.0"
pylint = "^2.17.2"
black = "^23.3.0"
isort = "^5.12.0"
ipython = "^8.12.0"

[tool.black]
line-length = 79
skip-string-normalization = true
docstring-style = "google"

[tool.pylint]
max-line-length = 79
disable = "C0103"

[tool.isort]
line_length = 88
multi_line_output = 3
include_trailing_comma = true
force_grid_wrap = 0
use_parentheses = true
sections="FUTURE,STDLIB,THIRDPARTY,FIRSTPARTY,LOCALFOLDER"
known_local_folder = ["app"]

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
